fee changes

+++ contracts/WorkSpaceFactoryLib.sol
@@ -9,7 +9,7 @@ struct FactoryState {
     mapping(address => uint32) currentIndex;
 
     uint32 amountOfWorkSpaces;
-    uint8 contractFee; //The fee can be maximum 100 and this is divided by 10 when its used as percentage. so 100 is 10%, 1 is 0.1%
+    uint16 contractFee; //The fee can be maximum 100 and this is divided by 10 when its used as percentage. so 100 is 10%, 1 is 0.1%
     bool disabled;
     address workSpaceLibraryAddress;
     address jobLibraryAddress;
@@ -55,7 +55,7 @@ library WorkSpaceFactoryLib {
     function getContractFee(FactoryState storage self)
         external
         view
-        returns (uint8)
+        returns (uint16)
     {
         return self.contractFee;
     }
